Add user registration before starting TodoList app
