+++ js/config/htmlElement.js
@@ -4,5 +4,14 @@ const $TODO_LIST = document.getElementById('todo-list');
 const $TODO_INPUT = document.getElementById('todo-input');
 const $TODO_COUNT = document.getElementsByClassName('count-container')[0];
 const $USER_LIST = document.getElementById('user-list');
+const $REGISTER_INPUT = document.getElementsByClassName('register-input')[0];
 
-export { $LOADING, $TODO_HEADER, $TODO_INPUT, $TODO_LIST, $TODO_COUNT, $USER_LIST };
+export {
+  $LOADING,
+  $TODO_HEADER,
+  $TODO_INPUT,
+  $TODO_LIST,
+  $TODO_COUNT,
+  $USER_LIST,
+  $REGISTER_INPUT
+};
